+++ src/Notifications/Entities/Simulation/Discount.php
@@ -6,6 +6,7 @@ namespace Paddle\SDK\Notifications\Entities\Simulation;
 
 use Paddle\SDK\FiltersUndefined;
 use Paddle\SDK\Notifications\Entities\DateTime;
+use Paddle\SDK\Notifications\Entities\Discount\DiscountMode;
 use Paddle\SDK\Notifications\Entities\Discount\DiscountStatus;
 use Paddle\SDK\Notifications\Entities\Discount\DiscountType;
 use Paddle\SDK\Notifications\Entities\Shared\CurrencyCode;
@@ -37,6 +38,8 @@ final class Discount implements SimulationEntity
         public readonly \DateTimeInterface|Undefined|null $expiresAt = new Undefined(),
         public readonly \DateTimeInterface|Undefined $createdAt = new Undefined(),
         public readonly \DateTimeInterface|Undefined $updatedAt = new Undefined(),
+        public readonly DiscountMode|Undefined|null $mode = new Undefined(),
+        public readonly string|Undefined|null $discountGroupId = new Undefined(),
     ) {
     }
 
@@ -60,6 +63,8 @@ final class Discount implements SimulationEntity
             expiresAt: self::optional($data, 'expires_at', fn ($value) => DateTime::from($value)),
             createdAt: self::optional($data, 'created_at', fn ($value) => DateTime::from($value)),
             updatedAt: self::optional($data, 'updated_at', fn ($value) => DateTime::from($value)),
+            mode: self::optional($data, 'mode', fn ($value) => DiscountMode::from($value)),
+            discountGroupId: self::optional($data, 'discount_group_id'),
         );
     }
 
@@ -72,6 +77,7 @@ final class Discount implements SimulationEntity
             'enabled_for_checkout' => $this->enabledForCheckout,
             'code' => $this->code,
             'type' => $this->type,
+            'mode' => $this->mode,
             'amount' => $this->amount,
             'currency_code' => $this->currencyCode,
             'recur' => $this->recur,
@@ -83,6 +89,7 @@ final class Discount implements SimulationEntity
             'expires_at' => $this->expiresAt,
             'created_at' => $this->createdAt,
             'updated_at' => $this->updatedAt,
+            'discount_group_id' => $this->discountGroupId,
         ]);
     }
 }