+++ src/Resources/DiscountGroups/Operations/CreateDiscountGroup.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Paddle\SDK\Resources\DiscountGroups\Operations;
+
+class CreateDiscountGroup implements \JsonSerializable
+{
+    public function __construct(
+        public readonly string $name,
+    ) {
+    }
+
+    public function jsonSerialize(): array
+    {
+        return [
+            'name' => $this->name,
+        ];
+    }
+}