+++ src/Resources/Discounts/Operations/CreateDiscount.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Paddle\SDK\Resources\Discounts\Operations;
 
+use Paddle\SDK\Entities\Discount\DiscountMode;
 use Paddle\SDK\Entities\Discount\DiscountType;
 use Paddle\SDK\Entities\Shared\CurrencyCode;
 use Paddle\SDK\Entities\Shared\CustomData;
@@ -30,6 +31,8 @@ class CreateDiscount implements \JsonSerializable
         public readonly array|Undefined|null $restrictTo = new Undefined(),
         public readonly string|Undefined|null $expiresAt = new Undefined(),
         public readonly CustomData|Undefined|null $customData = new Undefined(),
+        public readonly DiscountMode|Undefined $mode = new Undefined(),
+        public string|Undefined|null $discountGroupId = new Undefined(),
     ) {
     }
 
@@ -48,6 +51,8 @@ class CreateDiscount implements \JsonSerializable
             'restrict_to' => $this->restrictTo,
             'expires_at' => $this->expiresAt,
             'custom_data' => $this->customData,
+            'mode' => $this->mode,
+            'discount_group_id' => $this->discountGroupId,
         ]);
     }
 }